brd/server/requirements.txt
pandas==2.0.3
python-docx==0.8.11
Pillow==10.0.0
requests==2.31.0
markdown==3.4.4 